+++ perun/check/fast_check.py
@@ -24,13 +24,14 @@ def fast_check(baseline_profile, target_profile):
     )
 
 
-def exec_fast_check(baseline_profile, baseline_x_pts, abs_error):
+def exec_fast_check(uid, baseline_profile, baseline_x_pts, abs_error):
     """The function executes the classification of performance change between two profiles with
     using regression analysis. The type of the best model from the regressed profile, which
     contains the value absolute error, computed from the best models of both profile, is returned
     such as the degree of the changes.
 
-    :param dict baseline_profile: baseline against which we are checking the degradation
+    :param string uid: unique identifier of function for which we are creating the model
+    :param Profile baseline_profile: baseline against which we are checking the degradation
     :param np_array baseline_x_pts: the value absolute error computed from the linear models
         obtained from both profiles
     :param integer abs_error: values of the independent variables from both profiles
@@ -38,12 +39,25 @@ def exec_fast_check(baseline_profile, baseline_x_pts, abs_error):
     """
     # creating the new profile
     std_err_profile = copy.deepcopy(baseline_profile)
-    del std_err_profile['global']['models']
+    std_err_profile['models'].clear()
 
+    updated_data = {
+        'structure-unit-size': [],
+        'amount': []
+    }
     # executing the regression analysis
     for i, (x, y) in enumerate(zip(np.nditer(baseline_x_pts), np.nditer(abs_error))):
-        std_err_profile['global']['resources'][i]['structure-unit-size'] = x
-        std_err_profile['global']['resources'][i]['amount'] = y
+        updated_data['structure-unit-size'].append(x)
+        updated_data['amount'].append(y)
+    # Nasty hack, though it should work
+    std_err_profile._storage['resources'] = {
+        uid: updated_data
+    }
+    std_err_profile._storage['resource_type_map'] = {
+        uid: {
+            'uid': uid
+        }
+    }
 
     # Fixme: Extract of and per key
     regression_analysis_params = {