+++ CHANGELOG.rst
@@ -6,6 +6,17 @@ HEAD
 
 **To be included in next release**
 
+0.17 (2019-07-09)
+-----------------
+
+`commit: e6b1e88d766d93cdab4f114464df51114d6415a8`
+
+**Optimize profile format**
+
+  - make profile format more compact
+  - fix minor issue in fast check
+  - extract selected functions from query to profile object
+
 0.16.9-hotfix (2019-06-18)
 --------------------------
 
@@ -94,7 +105,7 @@ HEAD
 **Add kernel non-parametric regression**
 
   - fix minor issue in memory collector that manifests with gcc-5.5+ and Ubuntu 18.04+
-  - add three kernel non-parametrikjc regression models (see :ref:`_postprocessors-kernel-regression`)
+  - add three kernel non-parametrik regression models (see :ref:`_postprocessors-kernel-regression`)
   - fix minor issues in moving average and regressogram
 
 0.16.3 (2019-03-02)